+++ collective/notifications/config.py
@@ -1,0 +1,3 @@
+
+NOTIFICATION_KEY = 'collective.notifications'
+MAIN = '__notifications__'